+++ lodge/templates/navbar.html
@@ -13,7 +13,7 @@
     <!-- Collect the nav links, forms, and other content for toggling -->
     <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
       <ul class="nav navbar-nav navbar">
-        {% if request.path != '/' or '/error' or 'add_inquiry' or '/success' %}
+        {% if request.path != '/'%}
         <li><a href="{{'/'}}">Home</a></li>
         {%include "events/sidenav.html"%}
         </ul>
@@ -31,7 +31,7 @@
           <li><a href="{{'/wedding'}}">Weddings</a></li>
         </ul>
         </li>
-        <li><a href="{{'/newsletter'}}">Newsletter</a></li>
+        <li><a href="{{'/add_subscriber_to_newsletter_list'}}">Newsletter</a></li>
         <li><a href="#membership">Membership</a></li>
         <li><a href="#contact">Contact</a></li>
         </ul>
