+++ lodge/templates/user/newsletter.html
@@ -15,14 +15,12 @@
 			<div class="row">
 				<div class="col-sm"></div>
 				<div class="col-sm">
-				    <form class="form-signin">       
-				    	<h2 class="form-signin-heading">Please Enter your contact information</h2>
-				      	<input type="text" class="form-control" name="email" placeholder="Email" required="" autofocus=""  data-toggle="tooltip" title="please use this email for your login"/>
-				      	<br>
-				      	<input type="First Name" class="form-control" name="first name" placeholder="First Name" required=""  data-toggle="tooltip" title="use this username to login"/>
-				      	<br>
-				      	<input type="Last Name" class="form-control" name="last name" placeholder="Last Name" required=""  data-toggle="tooltip" title="use this username to login"/>
-				      	<br>
+					<form class="form-signin" method="post" action="{% url 'add_subscriber_to_newsletter_list' %}">
+				    	{{newsletter_form.as_p}}
+				    	{% csrf_token %}
+				    	<script src='https://www.google.com/recaptcha/api.js'></script>
+              			 <div class="g-recaptcha" data-sitekey="6Ld4YT8UAAAAAArKFbxAZ2lF8HqJWvR-Mo68EM_X"></div>
+    
 				      	<input type="radio" name="New Subscriber" value="subscriber">click here if you're a new subscriber
 				      	<br>
 				      	<br>
