+++ lodge/views/views.py
@@ -1,39 +1,40 @@
 from django.shortcuts import render
-from django.http import HttpResponse, HttpResponseRedirect, Http404
-from django.template import RequestContext
-from django.http import HttpResponse
+# from django.http import HttpResponse, HttpResponseRedirect, Http404
+# from django.template import RequestContext
+# from django.http import HttpResponse
 from django.shortcuts import redirect
 import operator
 from datetime import date
-from django import forms
-from lodge.forms.forms import *
-from django.db.models import Q
+# from django import forms
+# from lodge.forms.forms import *
+# from django.db.models import Q
 from lodge.models.models import *
 from django.conf import settings
-import braintree
-from django.core.mail import send_mail, BadHeaderError
-from django.core.mail import mail_admins
-from wsgiref.util import FileWrapper
+# import braintree
+# from django.core.mail import send_mail, BadHeaderError
+# from django.core.mail import mail_admins
+# from wsgiref.util import FileWrapper
+# from lodge.decorators.decorators import check_recaptcha
 
 
 
 
-braintree.Configuration.configure(
-    braintree.Environment.Sandbox,
-    'dwnqbjvjdjc8ysjp',
-    'kgzvv6qhfzq6dq5y',
-    '4e0b90ddfad27f835c658a2f76850504'
-)
+# braintree.Configuration.configure(
+#     braintree.Environment.Sandbox,
+#     'dwnqbjvjdjc8ysjp',
+#     'kgzvv6qhfzq6dq5y',
+#     '4e0b90ddfad27f835c658a2f76850504'
+# )
 
-TRANSACTION_SUCCESS_STATUSES = [
-    braintree.Transaction.Status.Authorized,
-    braintree.Transaction.Status.Authorizing,
-    braintree.Transaction.Status.Settled,
-    braintree.Transaction.Status.SettlementConfirmed,
-    braintree.Transaction.Status.SettlementPending,
-    braintree.Transaction.Status.Settling,
-    braintree.Transaction.Status.SubmittedForSettlement
-]
+# TRANSACTION_SUCCESS_STATUSES = [
+#     braintree.Transaction.Status.Authorized,
+#     braintree.Transaction.Status.Authorizing,
+#     braintree.Transaction.Status.Settled,
+#     braintree.Transaction.Status.SettlementConfirmed,
+#     braintree.Transaction.Status.SettlementPending,
+#     braintree.Transaction.Status.Settling,
+#     braintree.Transaction.Status.SubmittedForSettlement
+# ]
 # Create your views here.
 
 def index(request):
@@ -75,117 +76,122 @@ def error(request):
     # print("request", request.path)
     return render(request, template_name, {})
 
-def new_checkout(request):
-    client_token = braintree.ClientToken.generate()
-    template_name = 'paypal/new.html'
-    return render(request, template_name, {'client_token': client_token})
+# def new_checkout(request):
+#     client_token = braintree.ClientToken.generate()
+#     template_name = 'paypal/new.html'
+#     return render(request, template_name, {'client_token': client_token})
 
-def show_checkout(request, transaction_id):
-    transaction = braintree.Transaction.find(transaction_id)
-    template_name = "paypal/show.html"
-    result = {}
-    if transaction.status in TRANSACTION_SUCCESS_STATUSES:
-        result = {
-            'header': 'Sweet Success!',
-            'icon': 'success',
-            'message': 'Your test transaction has been successfully processed. See the Braintree API response and try again.'
-        }
-    else:
-        result = {
-            'header': 'Transaction Failed',
-            'icon': 'fail',
-            'message': 'Your test transaction has a status of ' + transaction.status + '. See the Braintree API response and try again.'
-        }
+# def show_checkout(request, transaction_id):
+#     transaction = braintree.Transaction.find(transaction_id)
+#     template_name = "paypal/show.html"
+#     result = {}
+#     if transaction.status in TRANSACTION_SUCCESS_STATUSES:
+#         result = {
+#             'header': 'Sweet Success!',
+#             'icon': 'success',
+#             'message': 'Your test transaction has been successfully processed. See the Braintree API response and try again.'
+#         }
+#     else:
+#         result = {
+#             'header': 'Transaction Failed',
+#             'icon': 'fail',
+#             'message': 'Your test transaction has a status of ' + transaction.status + '. See the Braintree API response and try again.'
+#         }
 
-    return render(request, template_name, {'transaction': transaction, 'result': result})
+#     return render(request, template_name, {'transaction': transaction, 'result': result})
 
-def create_checkout(request):
-	form = request.POST
-	data = form.dict()
-	result = braintree.Transaction.sale({
-	'amount': data['amount'],
-	'payment_method_nonce': data['payment_method_nonce'],
-	'options': {
-	"submit_for_settlement": True
-	}
-	})
-	if result.is_success or result.transaction:
-		return redirect('show_checkout', transaction_id=result.transaction.id)
-	else:
-		for x in result.errors.deep_errors: flash('Error: %s: %s' % (x.code, x.message))
-		return redirect('new_checkout')
-
-
-def archive(request):
-    template_name = 'events/allevents.html' 
-    events = Events.objects.order_by('event_date')
-    return render(request, template_name, {'events': events})
+# def create_checkout(request):
+# 	form = request.POST
+# 	data = form.dict()
+# 	result = braintree.Transaction.sale({
+# 	'amount': data['amount'],
+# 	'payment_method_nonce': data['payment_method_nonce'],
+# 	'options': {
+# 	"submit_for_settlement": True
+# 	}
+# 	})
+# 	if result.is_success or result.transaction:
+# 		return redirect('show_checkout', transaction_id=result.transaction.id)
+# 	else:
+# 		for x in result.errors.deep_errors: flash('Error: %s: %s' % (x.code, x.message))
+# 		return redirect('new_checkout')
 
 
-def search_keywords(request):
-    form_data = request.GET
-    iterable_form_data = form_data.dict()
-    search_box = iterable_form_data['Search']
-    template_name = 'events/allevents.html'
-    events_head = Events.objects.filter(event_description__contains=search_box)
-    if events_head.exists():
-        return render(request, template_name, {'events': events_head, 'search_box': search_box})
-    else: 
-        template_name = '404.html'
-        return render(request, template_name)
+# def archive(request):
+#     template_name = 'events/allevents.html' 
+#     events = Events.objects.order_by('event_date')
+#     return render(request, template_name, {'events': events})
 
-def filter_by_amount(request):
-    events = Events.objects.order_by('event_amount')
-    template_name = 'events/allevents.html'
-    return render(request, template_name, {'events': events})    
 
-def recent_events(request):
-    events = Events.objects.order_by('event_date')[:5]
-    template_name = 'events/allevents.html'
-    return render(request, template_name, {'events': events})    
+# def search_keywords(request):
+#     form_data = request.GET
+#     iterable_form_data = form_data.dict()
+#     search_box = iterable_form_data['Search']
+#     template_name = 'events/allevents.html'
+#     events_head = Events.objects.filter(event_description__contains=search_box)
+#     if events_head.exists():
+#         return render(request, template_name, {'events': events_head, 'search_box': search_box})
+#     else: 
+#         template_name = '404.html'
+#         return render(request, template_name)
 
-def get_this_event(request, event_id):
-    template_name = 'events/event.html'
-    event = Events.objects.get(pk= event_id)
-    return render(request, template_name, {'event': event})
+# def filter_by_amount(request):
+#     events = Events.objects.order_by('event_amount')
+#     template_name = 'events/allevents.html'
+#     return render(request, template_name, {'events': events})    
 
-def add_inquiry(request):
-    if request.method == 'GET':
-        template_name = 'user/inquiry.html'
-        inquiry_form = InquiryForm()
-        return render(request, template_name, {'inquiry_form': inquiry_form})
+# def recent_events(request):
+#     events = Events.objects.order_by('event_date')[:5]
+#     template_name = 'events/allevents.html'
+#     return render(request, template_name, {'events': events})    
 
-    elif request.method == 'POST':
-        form_data = request.POST
-        inquiry_form = InquiryForm()
+# def get_this_event(request, event_id):
+#     template_name = 'events/event.html'
+#     event = Events.objects.get(pk= event_id)
+#     return render(request, template_name, {'event': event})
+
+# @check_recaptcha
+# def add_inquiry(request):
+#     if request.method == 'GET':
+#         template_name = 'user/inquiry.html'
+#         inquiry_form = InquiryForm()
+#         return render(request, template_name, {'inquiry_form': inquiry_form})
+
+#     elif request.method == 'POST':
+#         form_data = request.POST
+#         inquiry_form = InquiryForm()
 
         
-        i = Inquiry(
-            inquirer_first_name = form_data['inquirer_first_name'],
-            inquirer_last_name = form_data['inquirer_last_name'],
-            inquiry_title  = form_data['inquiry_title'],
-            inquiry_content = form_data['inquiry_content'],
-            inquiry_email_address = form_data['inquiry_email_address'],
-        ) 
+#         i = Inquiry(
+#             inquirer_first_name = form_data['inquirer_first_name'],
+#             inquirer_last_name = form_data['inquirer_last_name'],
+#             inquiry_title  = form_data['inquiry_title'],
+#             inquiry_content = form_data['inquiry_content'],
+#             inquiry_email_address = form_data['inquiry_email_address'],
+#         ) 
 
-        i.save()
+#         i.save()
         
-        botcheck = 'yes'
-        subject = "Elk Lodge Inquiry"
-        message = "This is a confirmation of your recent inquiry to the Elk Lodge"
-        from_email = "imfake@email.com"
-        recipient = form_data['inquiry_email_address']
-        recipient_list = [recipient]
-        if subject and message and recipient_list and botcheck == 'yes':
-            try:
-                send_mail(subject, message, from_email, recipient_list)
-                mail_admins(subject, message, fail_silently=False, connection=None, html_message=None)
-            except BadHeaderError:
-                return HttpResponse('Invalid header found.')
-            return HttpResponseRedirect('/success')
-        else:
-            return HttpResponse('Make sure all fields are entered and valid.')
-            return HttpResponseRedirect('/error')
+#         # botcheck = 'yes'
+#         subject = "Elk Lodge Inquiry"
+#         message = "This is a confirmation of your recent inquiry to the Elk Lodge"
+#         from_email = "imfake@email.com"
+#         recipient = form_data['inquiry_email_address']
+#         recipient_list = [recipient]
+
+#         if i.is_valid() and request.recaptcha_is_valid:
+#         # if subject and message and recipient_list and botcheck == 'yes':
+#             try:
+#                 send_mail(subject, message, from_email, recipient_list)
+#                 mail_admins(subject, message, fail_silently=False, connection=None, html_message=None)
+#             except BadHeaderError:
+#                 return HttpResponse('Invalid header found.')
+#             return HttpResponseRedirect('/success')
+#         else:
+#             return HttpResponse('Make sure all fields are entered and valid.')
+#             return HttpResponseRedirect('/error')
+
+
