+++ tmp/email_messages/20180107-012930-4581994112.log
@@ -1,0 +1,12 @@
+Content-Type: text/plain; charset="utf-8"
+MIME-Version: 1.0
+Content-Transfer-Encoding: 7bit
+Subject: Elk Lodge Newsletter Subscription request
+From: imfake@email.com
+To: qw3e23@sge.com
+Date: Sun, 07 Jan 2018 01:29:30 -0000
+Message-ID: 
+ <20180107012930.36321.82636@1.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.7.6.8.0.2.d.e.0.0.7.1.0.0.6.2.ip6.arpa>
+
+This is a confirmation of your Newsletter subscription request
+-------------------------------------------------------------------------------
