+++ tmp/email_messages/20180107-013809-4539197928.log
@@ -1,0 +1,12 @@
+Content-Type: text/plain; charset="utf-8"
+MIME-Version: 1.0
+Content-Transfer-Encoding: 7bit
+Subject: Elk Lodge Inquiry
+From: imfakeinquiry@email.com
+To: adgraerg@l.com
+Date: Sun, 07 Jan 2018 01:38:09 -0000
+Message-ID: 
+ <20180107013809.36422.45715@1.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.7.6.8.0.2.d.e.0.0.7.1.0.0.6.2.ip6.arpa>
+
+This is a confirmation of your recent inquiry to the Elk Lodge
+-------------------------------------------------------------------------------
