+++ tmp/email_messages/20180107-014132-4423143096.log
@@ -1,0 +1,12 @@
+Content-Type: text/plain; charset="utf-8"
+MIME-Version: 1.0
+Content-Transfer-Encoding: 7bit
+Subject: Elk Lodge Newsletter Subscription request
+From: imfakenewslettertest@email.com
+To: erateqrt@gmail.com
+Date: Sun, 07 Jan 2018 01:41:32 -0000
+Message-ID: 
+ <20180107014132.36444.55852@1.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.7.6.8.0.2.d.e.0.0.7.1.0.0.6.2.ip6.arpa>
+
+This is a confirmation of your Newsletter subscription request
+-------------------------------------------------------------------------------
